+++ src/components/User/Accordion/styles.module.css
@@ -1,0 +1,22 @@
+.root>div {
+  background-color: #1a1a1a !important;
+}
+
+.root h2>span {
+  background-color: #272727 !important;
+}
+
+.root>div,
+.root h2>span {
+  color: white !important;
+}
+
+.root h2>span:hover {
+  cursor: pointer;
+}
+
+/* стрелка выпадающего листа */
+.root h2>span::after {
+  background-image: var(--bs-accordion-btn-icon-custom) !important;
+  transition: var(--bs-accordion-btn-icon-custom-transition);
+}
