+++ src/components/User/Cabinet/Cabinet.tsx
@@ -1,4 +1,5 @@
 import Avatar from "../Avatar/Avatar"
+import Accordion from "../Accordion/Accordion"
 import styles from "./styles.module.css"
 import classNames from "classnames"
 
@@ -17,44 +18,10 @@ export default function Content({ user }: Props) {
         <Avatar userPhoto={user.photo} />
       </div>
 
-      <div className="accordion" id={styles.accordion}>
-
-        <div className="accordion-item">
-          <h2 className="accordion-header" onClick={(event) => collapser(event)}>
-            <span className="accordion-button">
-              пользователь
-            </span>
-          </h2>
-          <div className="accordion-collapse">
-            <div className="accordion-body">
-              <p>email: {user.email}</p>
-              <p>ранг: {user.rank}</p>
-              <p>должность: {user.position || "не указана"}</p>
-            </div>
-          </div>
-        </div>
-
-
-        <div className="accordion-item">
-          <h2 className="accordion-header" onClick={(event) => collapser(event)}>
-            <span className="accordion-button collapsed">
-              доп. информация
-            </span>
-          </h2>
-          <div className="accordion-collapse collapse">
-            <div className="accordion-body">
-              lorem ipsum
-            </div>
-          </div>
-        </div>
-
+      <div>
+        <Accordion user={user} />
       </div>
 
     </div>
   </div>
-}
-
-function collapser(event: React.MouseEvent<HTMLHeadingElement, MouseEvent>) {
-  event.currentTarget.firstElementChild?.classList.toggle("collapsed")
-  event.currentTarget.nextElementSibling?.classList.toggle("collapse")
 }
