+++ src/components/User/Cabinet/styles.module.css
@@ -24,28 +24,3 @@ p {
 .content>div:nth-child(2) {
   flex: 5 500px;
 }
-
-
-/* аккордеон */
-#accordion>div {
-  background-color: #1a1a1a !important;
-}
-
-#accordion h2>span {
-  background-color: #272727 !important;
-}
-
-#accordion>div,
-#accordion h2>span {
-  color: white !important;
-}
-
-#accordion h2>span:hover {
-  cursor: pointer;
-}
-
-/* стрелка выпадающего листа */
-#accordion h2>span::after {
-  background-image: var(--bs-accordion-btn-icon-custom) !important;
-  transition: var(--bs-accordion-btn-icon-custom-transition);
-}
